+++ app/controllers/profiles_controller.rb
@@ -53,7 +53,7 @@ class ProfilesController < ApplicationController
 
     respond_to do |format|
       format.js { render :nothing => true, :status => 200 }
-      format.html {
+      format.any {
         flash[:notice] = I18n.t 'profiles.update.updated'
         if current_user.getting_started?
           redirect_to getting_started_path